Split quarkus-cache into extension and API library
Created quarkus-cache-api module with annotations and interfaces
following the same pattern as quarkus-scheduler-api.

Changes:
- Created extensions/cache/api with 13 API classes
- Moved annotations, interfaces, and key classes from runtime to API
- Updated runtime module to depend on API
- Updated deployment module to depend on API
- Added cache-api to BOM
- Moved UndefinedCacheKeyGenerator from runtime package to API

This allows library projects to use cache annotations without
pulling in the full cache extension, reducing dependency footprint.

Fixes #56189
