+++ bom/application/pom.xml
@@ -3537,6 +3537,11 @@
             </dependency>
             <dependency>
                 <groupId>io.quarkus</groupId>
+                <artifactId>quarkus-cache-api</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>io.quarkus</groupId>
                 <artifactId>quarkus-cache</artifactId>
                 <version>${project.version}</version>
             </dependency>
