+++ extensions/cache/api/src/main/java/io/quarkus/cache/CacheManager.java
