+++ extensions/cache/api/src/main/java/io/quarkus/cache/CacheName.java
