+++ extensions/cache/pom.xml
@@ -16,6 +16,7 @@
     <packaging>pom</packaging>
 
     <modules>
+        <module>api</module>
         <module>deployment</module>
         <module>deployment-spi</module>
         <module>kotlin</module>
