+++ extensions/cache/runtime/src/main/java/io/quarkus/cache/runtime/CacheInterceptor.java
@@ -30,6 +30,7 @@ import io.quarkus.cache.CacheKey;
 import io.quarkus.cache.CacheKeyGenerator;
 import io.quarkus.cache.CacheManager;
 import io.quarkus.cache.CompositeCacheKey;
+import io.quarkus.cache.UndefinedCacheKeyGenerator;
 import io.smallrye.mutiny.Uni;
 
 public abstract class CacheInterceptor {
